Fix asdf system definition
The system definition assumed that the path to Snark was configured
using :TREE, not :DIRECTORY (or, more exactly, that both the root
directory and the `src' directory are in the source registry).  This
change fixes this assumption by moving all systems definition files
into the root directory.  A better solution would probably be to have
just one system definition with several components.